requirements.txt
# backend/requirements.txt

Django>=5.0
djangorestframework
gunicorn
django-cors-headers
waitress==2.1.1

